+++ eitprocessing/plotting/pixelmap.py
